backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
anthropic==0.40.0
langchain==0.3.7
langchain-community==0.3.7
langchain-anthropic==0.3.0
langgraph==0.2.50
faiss-cpu==1.13.2
sentence-transformers==3.3.1
beautifulsoup4==4.12.3
requests==2.32.3
numpy
reportlab==4.2.5
